Split up the lambdas
