+++ aws/src/main/python/budget_instance_stopper/budgeted_instance_stopper.yaml
@@ -2,11 +2,11 @@ AWSTemplateFormatVersion: '2010-09-09'
 Transform: 'AWS::Serverless-2016-10-31'
 Description: An AWS Serverless Specification template describing your function.
 Resources:
-  budgetnotifier:
+  budgetedinstancestopper:
     Type: 'AWS::Serverless::Function'
     Properties:
       Handler: lambda_function.lambda_handler
-      Runtime: python3.8
+      Runtime: python3.9
       CodeUri: .
       Description: ''
       MemorySize: 128
@@ -14,4 +14,6 @@ Resources:
       Role: 'arn:aws:iam::340032650202:role/BeamLambdaExecution'
       Environment:
         Variables:
+          EXCLUDED_INSTANCES: '[''i-0d7356ff7f3927e69'',''i-0b31829f99cc9a7ce'',''i-01dda95dee4c4714f'']'
+          MONITORED_REGIONS: '[''us-east-1'',''us-east-2'',''us-west-1'',''us-west-2'']'
           SLACK_HOOK: SENSITIVE_SO_REDACTED
